% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.binding_docs_and_data_eea.R
\name{all_link_df}
\alias{all_link_df}
\title{Une ambos dataframes para cada firma.}
\usage{
all_link_df(all_docs_corr, all_data_corr)
}
\arguments{
\item{all_docs_corr}{dataframe para documentos}

\item{all_data_corr}{dataframe para datos}
}
\value{
devuelve un dataframe
}
\description{
Une ambos dataframes para cada firma.
}
\examples{
all_link_df(all_docs_corr,all_data_corr)
}
