% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.1.all_eea_data_2001_2020.R
\name{datos_peru}
\alias{datos_peru}
\title{Datos peru}
\usage{
datos_peru(datos_bloue1, datos_bloue2, datos_bloue3, datos_bloue4, datos_dap1)
}
\arguments{
\item{datos_bloue1}{datos bloque 1, o periodo 1}

\item{datos_bloue2}{datos bloque 2}

\item{datos_bloue3}{datos bloque 3}

\item{datos_bloue4}{datos bloque 4}

\item{datos_dap1}{datos generales}
}
\value{
dataframe para peru
}
\description{
Datos peru
}
\examples{
datos_peru(datos_bloue1,datos_bloue2,datos_bloue3,datos_bloue4,datos_dap1)
}
