% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07. download_rawdata_eea.R
\name{download_rawdata}
\alias{download_rawdata}
\title{Descarga los datos}
\usage{
download_rawdata(eea_data_links_variables_df_manual)
}
\arguments{
\item{eea_data_links_variables_df_manual}{El manual con los capitulos rellenados previamente}
}
\value{
Devuelve un raw data frame con todas las variables, aun no se elige las variables
}
\description{
Descarga los datos
}
\examples{
download_rawdata(eea_data_links_variables_df_manual)
}
