% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.identificando_variables_firms_eea.R
\name{id_variables}
\alias{id_variables}
\title{Opcional, solo cuando sale el 2021 en adelante, para completar manualmente el excel de capitulos de las variables}
\usage{
id_variables(temp_all_data)
}
\arguments{
\item{temp_all_data}{los enlaces unificados}
}
\value{
nada, solo te provee un chrome que se actualiza cada cierto intervalo para que rellenes el manual
}
\description{
Opcional, solo cuando sale el 2021 en adelante, para completar manualmente el excel de capitulos de las variables
}
\examples{
id_variables(temp_all_data)
}
