% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.rename_sector_firms_level_eea.R
\name{processing_sector_name}
\alias{processing_sector_name}
\title{Renombra, formatea o uniformiza los nombres}
\usage{
processing_sector_name(
  only_firms_doc_link,
  eea_data_links,
  correspondencia_datos,
  correspondencia_docs,
  tildes
)
}
\arguments{
\item{only_firms_doc_link}{dataframe de enlaces de doc}

\item{eea_data_links}{dataframe de enlace de datos}

\item{correspondencia_datos}{excel de correspondencia datos}

\item{correspondencia_docs}{excel de correspondencia docs}

\item{tildes}{caracteres especiales, vocales mayusculas tildadas}
}
\value{
datafrme renombrado uniformemente
}
\description{
Renombra, formatea o uniformiza los nombres
}
\examples{
processing_sector_name(only_firms_doc_link,eea_data_links,correspondencia_datos,correspondencia_docs,tildes)
}
