% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.scraping_link_eea.R
\name{scraping_eea_datalink_function}
\alias{scraping_eea_datalink_function}
\title{Scraper de enlaces de datos}
\usage{
scraping_eea_datalink_function(start_date)
}
\arguments{
\item{start_date}{Periodo de inicio}
}
\value{
Df de enlaces
}
\description{
Scraper de enlaces de datos
}
\examples{
scraping_eea_datalink_function(start_date)
}
