% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08. select_rawdata_eea_2000_2011.R
\name{select_2000_2011}
\alias{select_2000_2011}
\title{Selecciona datos de 2000 a 2011, por semejanza}
\usage{
select_2000_2011(eea_data_links_variables_df_manual, eea_raw_data)
}
\arguments{
\item{eea_data_links_variables_df_manual}{excel de capitulos}

\item{eea_raw_data}{datos brutos}
}
\value{
lista de dataframe, una para cada variable
}
\description{
Selecciona datos de 2000 a 2011, por semejanza
}
\examples{
select_2000_2011(eea_data_links_variables_df_manual,eea_raw_data)
}
