% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09.select_raw_data_2012_1.R
\name{select_2012_1}
\alias{select_2012_1}
\title{Selecciona datos solo de 2012, por diferencia del resto}
\usage{
select_2012_1(eea_data_links_variables_df_manual, eea_raw_data)
}
\arguments{
\item{eea_data_links_variables_df_manual}{excel de capitulos}

\item{eea_raw_data}{datos brutos}
}
\value{
lista de dataframe, una para cada variable
}
\description{
Selecciona datos solo de 2012, por diferencia del resto
}
\examples{
select_2012_1(eea_data_links_variables_df_manual,eea_raw_data)
}
