% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10.select_raw_data_2012_2_2019.R
\name{select_2012_2019}
\alias{select_2012_2019}
\title{Selecciona datos de 2012 a 2019, tambien por semejanza, no es admisible tilde}
\usage{
select_2012_2019(eea_data_links_variables_df_manual, eea_raw_data)
}
\arguments{
\item{eea_data_links_variables_df_manual}{excel de capitulos}

\item{eea_raw_data}{datos brutos}
}
\value{
lista de dataframe, una para cada variable
}
\description{
Selecciona datos de 2012 a 2019, tambien por semejanza, no es admisible tilde
}
\examples{
select_2012_2019(eea_data_links_variables_df_manual,eea_raw_data)
}
