% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.select_raw_data_2020.R
\name{select_2020}
\alias{select_2020}
\title{Selecciona los datos de 2020}
\usage{
select_2020(eea_data_links_variables_df_manual, eea_raw_data)
}
\arguments{
\item{eea_data_links_variables_df_manual}{excel de capitulos}

\item{eea_raw_data}{datos brutos}
}
\value{
lista de dataframe, una para cada variable
}
\description{
Selecciona los datos de 2020
}
\examples{
select_2020(eea_data_links_variables_df_manual,eea_raw_data)
}
