% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12.select_raw_todos_data_2020.R
\name{select_cap1}
\alias{select_cap1}
\title{Selecciona los datos de todos, capitulo 1}
\usage{
select_cap1(eea_data_links_variables_df_manual, eea_raw_data, tildes)
}
\arguments{
\item{eea_data_links_variables_df_manual}{excel de capitulos}

\item{eea_raw_data}{datos brutos}

\item{tildes}{vocales en tildes}
}
\value{
dataframe con generalidades de cada firma
}
\description{
Selecciona los datos de todos, capitulo 1
}
\examples{
select_cap1(eea_data_links_variables_df_manual,eea_raw_data)
}
