% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.select_firms_level_eea.R
\name{select_only_firm}
\alias{select_only_firm}
\title{Selecciona solo firmas}
\usage{
select_only_firm(
  establecimientos_xlsx,
  start_date,
  end_date,
  data_links,
  doc_links
)
}
\arguments{
\item{establecimientos_xlsx}{Archivo de establecimientos para eliminar}

\item{start_date}{periodo de inicio}

\item{end_date}{periodo final}

\item{data_links}{link de datos}

\item{doc_links}{link de documentacion}
}
\value{
dataframe de documentacion solo para establecimientos
}
\description{
Selecciona solo firmas
}
\examples{
select_only_firm(establecimientos_xlsx,start_date,end_date,data_links,doc_links)
}
